fixed various warnings
git-svn-id: https://svn.salilab.org/imp/trunk@13030 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/isd/include/internal/cg_eigen.h
@@ -44,7 +44,7 @@ public:
      */
     void set_A(const MatrixXd& A)
     {
-        double M = A.rows();
+        unsigned M = A.rows();
         IMP_USAGE_CHECK(A.cols() == M, "must provide spd matrix!");
         M_ = M;
         A_ = A;
